% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CausalStump}
\alias{CausalStump}
\title{Fit a Gaussian process or Student-t process to the treatment and control surface.}
\usage{
CausalStump(y, X, z, w, pscore, kernelfun = "SE", myoptim = "Nadam",
  maxiter = 5000, minimumRMSE = FALSE, tol = 1e-04, prior = FALSE,
  nu = 200, nsampling = 5000, learning_rate = 0.01, beta1 = 0.9,
  beta2 = 0.999, momentum = 0)
}
\arguments{
\item{y}{A vector}

\item{X}{A data.frame}

\item{z}{A vector}

\item{w}{A vector (optional)}

\item{pscore}{A vector (optional)}

\item{kernelfun}{A string (default: SE)}

\item{myoptim}{A string (default: Gradient Descent -- GD)}

\item{maxiter}{(default: 5000)}

\item{tol}{(default: 1e-4)}

\item{prior}{A logic statement (default: FALSE)}

\item{nu}{A value (default: 200)}

\item{nsampling}{A value (default: 5000) number of samples for prediction}

\item{learning_rate}{(default: 0.01)}

\item{beta1}{(default: 0.9)}

\item{beta2}{(default: 0.999)}

\item{momentum}{(default: 0.0)}
}
\value{
The function returns the fitted process as a CausalStump class object
}
\description{
Fit a Gaussian process or Student-t process to the treatment and control surface.
}
\examples{
#Generate data
n = 120
Z = rbinom(n, 1, 0.3)
X0 = runif(n-sum(Z), min = 20, max = 40)
X = data.frame(matrix(NaN,n,1))
X[Z==1,] = X1; X[Z==0,] = X0
y0_true = as.matrix(72 + 3 * sqrt(X))
y1_true = as.matrix(90 + exp(0.06 * X))
Y0 = rnorm(n, mean = y0_true, sd = 1)
Y1 = rnorm(n, mean = y1_true, sd = 1)
Y = Y0*(1-Z) + Y1*Z
mystump <- CausalStump(Y,X,Z)
}
